requirements.txt
altgraph==0.17.4
jpype1 @ file:///home/conda/feedstock_root/build_artifacts/jpype1_1738063513548/work
modulegraph==0.19.6
numpy==2.2.6
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.2
py_ltl==0.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
